# static/openapi.yaml
openapi: 3.1.0
info:
  title: Finault Gateway API
  version: 1.0.0
  description: |
    Comprehensive API for invoice processing, cost allocation, financial reconciliation,
    and intelligent expense management. Supports 47+ invoice formats, multi-cloud ERP integration,
    cryptographic proof verification, and AI-powered savings intelligence.
  contact:
    name: Finault Support
    url: https://finault.com
    email: support@finault.com
  license:
    name: Proprietary
    url: https://finault.com/license
servers:
  - url: https://api.finault.com
    description: Production environment
  - url: http://localhost:8787
    description: Development environment
security:
  - bearerAuth: []
  - apiKeyHeader: []
  - finaultKeyHeader: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authentication
    finaultKeyHeader:
      type: apiKey
      in: header
      name: x-finault-key
      description: Finault API key for authentication
...
paths:
...
  /v1/audit/export:
    get:
      tags:
        - Audit & Compliance
      summary: Export audit trail
      description: Export audit trail for SOX/SOC2 compliance
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, json]
      responses:
        '200':
          description: Audit trail export
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string

  # AI Proxy
...